ボトムナビゲーションバー追加（追加に伴い，main.dartいじりまくりました．clockページは，clock_home.dartに入ってます）
