+++ lib/quiz_home.dart
@@ -1,0 +1,18 @@
+import 'package:flutter/material.dart';
+
+
+class QuizHome extends StatelessWidget {
+  const QuizHome({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      appBar: AppBar(
+        title: Text('メモクロ'),
+        backgroundColor: Colors.purple,
+      ),
+      body: null,
+    );
+  }
+}
+
