+++ test/widget_test.dart
@@ -9,6 +9,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:memorization_and_clock/main.dart';
+import 'package:memorization_and_clock/route.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
